blpx: Add advanced expression syntax
Added support for advanced expression syntax, which compiles to a
standard set of closure names.

+++ blueprintcompiler/lsp.py
@@ -78,7 +78,7 @@ class OpenFile:
         self.diagnostics: list[CompileError] = []
         try:
             self.tokens = tokenizer.tokenize(self.text)
-            self.ast, errors, warnings = parser.parse(self.tokens)
+            self.ast, errors, warnings = parser.parse(self.tokens, blpx=True)
             self.diagnostics += warnings
             if errors is not None:
                 self.diagnostics += errors.errors
@@ -443,16 +443,19 @@ class LanguageServer:
 
         def action_to_json(action: CodeAction, diagnostic: CompileError):
             assert diagnostic.range is not None
-            edits = [
-                {
-                    "range": (
-                        action.edit_range.to_json()
-                        if action.edit_range
-                        else diagnostic.range.to_json()
-                    ),
-                    "newText": action.replace_with,
-                }
-            ]
+            edits = []
+
+            if action.replace_with is not None:
+                edits.append(
+                    {
+                        "range": (
+                            action.edit_range.to_json()
+                            if action.edit_range
+                            else diagnostic.range.to_json()
+                        ),
+                        "newText": action.replace_with,
+                    }
+                )
 
             for edit in action.additional_edits or []:
                 edits.append(edit.to_json())
